Refactor RemoveJsonPropertyDiagnosticCodeActionProvider into a more sharable location.
